Expose Sleep and explicitly stop timer
This commit replaces timeSleeper with Sleep.

Previously, if the context finishes early, timeSleeper returns the
context's error and ignore the timer.
This is correct, but the GC will only be able to collect
the timer after the timer expires.
This commit explicitly stop the timer, so that the GC can claim it
immediately.