all: enable modules
This adds a go.mod and go.sum file to gax.

Unfortunately, gax-go already has a v2.0.0 tag. This
means that the module must be a v2 module. So, per
[1] we create a v2/ subdirectory, copy all the v1 code
there, and mark the go.mod module declaration with
/v2.

However, there's one more complication: users are
directly exposed to exported types in this package
such as CallOption. google-cloud-go will rely on v2
of gax, but a user's codebase is still going to have
the non-v2/ gax import paths.

We could:

- Make google-cloud-go rely on v2 and simultaneously
create a v2 of google-cloud-go. All existing users of
google-cloud-go continue using v1 of gax and v1 of
google-cloud-go. All folks migrating to v2 of
google-cloud-go must go update their gax import paths
to use /v2. (this is obviously infeasible)
- Make all gax v1 types aliases of gax v2 types, and
all v1 functions shadows of v2 functions. That way,
google-cloud-go can rely on v2 but a user can pass
either a v1 or v2 gax type.

This change will be tagged as v1.0.1 as well as
v2.0.1. The versions v1.0.0 and v2.0.0 will be unusable.

1: https://github.com/golang/go/wiki/Modules#releasing-modules-v2-or-higher