chore(all): update to 'go 1.17' to enable module graph pruning
Updating the 'go' directive to 'go 1.17' will enable module graph
pruning (https://go.dev/ref/mod#graph-pruning), which will allow users
who depend on these modules to prune out irrelevant transitive
dependencies.

Note that updating to 'go 1.17' *will not* break existing users on
versions older than 1.17. (The 'go' version is the version required
for full fidelity — in this case, graph pruning — but older versions
of the Go toolchain will continue to make a best-effort attempt to
compile packages and will not report an error if that attempt
succeeds, as it will in this case.)